requirements.txt
pandas == 2.2.3
matplotlib == 3.9.2
seaborn == 0.13.2
numpy
scikit-learn
pyyaml
pgeocode == 0.5.0
flask
wandb